+++ changelog.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file. The format 
 - New split() method to Font class which works like wrap() but provides a list of strings instead of a single formatted one.
 - New runGarbageCollection terminal command that can be used to reclaim heap memory at runtime.
 - New terminal commands listMonitors and listSpeakers for exposing available peripheral devices to developers.
+- targetDelta and trueDelta to Camera.update() method for smooth interpolation if needed.
 
 ### Changed
 - Fixed bug that wouldn't update the value of virtual gamepads when Input.setVirtualGamepadInput() was called.